requirements.txt
discord.py
yt-dlp
python-dotenv
pynacl
google-generativeai
